chore: address comments and improve test coverage (6)
